feat(docker): provide scratch docker file
It's a lot easier for me to just have the local Docker daemon restart
the application than remembering to do it myself. This provides a
Dockerfile and docker-compose.yml for ease of restarting the
application, opening the port, mapping the config.toml, etc.

DCO-1.1 Signed-off-by: Patrick Wagstrom <160672+pridkett@users.noreply.github.com>